arrange files in components
