+++ src/components/map/mapMarker/index.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import mapboxgl from 'mapbox-gl';
 
-import { useMapbox } from '../../../context/mapContext';
+import { useMapbox } from '../mainMap/context';
 import './index.css';
 /*
   Add marker on map
